requirements.txt
fastmcp
langgraph
langchain-openai
langchain-mcp-adapters
pydantic
python-dotenv